#166057619: Implement Pitch Feedback

+++ server/migrations/tables.js
@@ -17,18 +17,18 @@ const accounts = `CREATE TABLE IF NOT EXISTS accounts(
     owner INT NOT NULL,
     type VARCHAR NOT NULL,
     status VARCHAR NOT NULL DEFAULT 'draft',
-    balance FLOAT NOT NULL
+    balance NUMERIC(15, 2) NOT NULL
 );`;
 
 const transactions = `CREATE TABLE IF NOT EXISTS transactions(
     id SERIAL PRIMARY KEY,
     createdOn TIMESTAMP DEFAULT NOW(),
     type VARCHAR NOT NULL,
-    amount FLOAT NOT NULL,
+    amount NUMERIC(15, 2) NOT NULL,
     accountNumber INT NOT NULL,
     cashier INT NOT NULL,
-    oldBalance FLOAT NOT NULL,
-    newBalance FLOAT NOT NULL
+    oldBalance NUMERIC(15, 2) NOT NULL,
+    newBalance NUMERIC(15, 2) NOT NULL
 );`;
 
 db.query(users, (error) => {
